requirements.txt
tabulate
pandas
requests
numpy
sqlalchemy
psycopg2-binary
python-dotenv
serpapi